fixed one typo

+++ c_src/phase.c
@@ -519,8 +519,8 @@ int mc_set_init(Phase * const p){
     poly_order[poly_i-i]=poly_i;
   }
 
-  for(unsigned int index=0;index<num_long_chain;index++){
-    if(poly_order[index]!=index){
+  for(int index=0;index<num_long_chain;index++){
+    if(poly_order[index]!=(unsigned int) index){
       exchange_polymer(p,poly_order[index],index);  
     }
   }
